# web/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/campaigns/{campaignId}/ask/conversations": {
      "get": {
        "operationId": "AskController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedAskConversationDto"
                }
              }
            }
          }
        },
        "tags": [
          "Ask"
        ]
      },
      "post": {
        "operationId": "AskController_create",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskConversationDto"
                }
              }
            }
          },
          "403": {
            "description": "You must be part of this campaign to talk to the Bard."
          }
        },
        "tags": [
          "Ask"
        ]
      }
    },
    "/api/v1/campaigns/{campaignId}/ask/conversations/{conversationId}": {
      "get": {
        "operationId": "AskController_detail",
        "parameters": [
          {
            "name": "conversationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskConversationDetailDto"
                }
              }
            }
          },
          "403": {
            "description": "This conversation is private."
          },
          "404": {
            "description": "No such conversation in this campaign."
          }
        },
        "tags": [
          "Ask"
        ]
      },
      "patch": {
        "operationId": "AskController_update",
        "parameters": [
          {
            "name": "conversationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskConversationPatchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Conversation renamed and/or shared."
          },
          "400": {
            "description": "Empty or oversized title."
          },
          "403": {
            "description": "Only the author can change this conversation."
          }
        },
        "tags": [
          "Ask"
        ]
      },
      "delete": {
        "operationId": "AskController_remove",
        "parameters": [
          {
            "name": "conversationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Only the author can delete this conversation."
          },
          "404": {
            "description": "No such conversation in this campaign."
          }
        },
        "tags": [
          "Ask"
        ]
      }
    },
    "/api/v1/campaigns/{campaignId}/ask/conversations/{conversationId}/messages": {
      "post": {
        "operationId": "AskController_askQuestion",
        "parameters": [
          {
            "name": "conversationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskQuestionBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskAnswerDto"
                }
              }
            }
          },
          "400": {
            "description": "Empty or oversized question."
          },
          "403": {
            "description": "Only the author can ask in this conversation."
          },
          "404": {
            "description": "No such conversation in this campaign."
          },
          "409": {
            "description": "The Bard is still answering in this conversation."
          }
        },
        "tags": [
          "Ask"
        ]
      }
    },
...
    "/api/v1/campaigns/{campaignId}/{entityType}": {
      "post": {
        "operationId": "EntityCrudController_create",
        "parameters": [
          {
            "name": "entityType",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "npcs",
                "locations",
                "factions",
                "quests",
                "inventory",
                "artifacts",
                "bestiary",
                "timeline"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityMutationResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid entity type or field values."
          },
          "403": {
            "description": "You must be part of this campaign to change it."
          },
          "409": {
            "description": "An entity with the same natural key already exists."
          }
        },
        "tags": [
          "EntityCrud"
        ]
      }
    },
    "/api/v1/campaigns/{campaignId}/{entityType}/{shortId}": {
      "patch": {
        "operationId": "EntityCrudController_update",
        "parameters": [
          {
            "name": "entityType",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "npcs",
                "locations",
                "factions",
                "quests",
                "inventory",
                "artifacts",
                "bestiary",
                "timeline"
              ],
              "type": "string"
            }
          },
          {
            "name": "shortId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityMutationResultDto"
                }
              }
            }
          },
          "403": {
            "description": "You must be part of this campaign to change it."
          },
          "404": {
            "description": "No entity with that short id in this campaign."
          },
          "409": {
            "description": "The new name is already taken by another entity."
          }
        },
        "tags": [
          "EntityCrud"
        ]
      },
      "delete": {
        "operationId": "EntityCrudController_remove",
        "parameters": [
          {
            "name": "entityType",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "npcs",
                "locations",
                "factions",
                "quests",
                "inventory",
                "artifacts",
                "bestiary",
                "timeline"
              ],
              "type": "string"
            }
          },
          {
            "name": "shortId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityDeleteResultDto"
                }
              }
            }
          },
          "400": {
            "description": "This entity cannot be deleted (e.g. the Party faction)."
          },
          "403": {
            "description": "You must be part of this campaign to change it."
          },
          "404": {
            "description": "No entity with that short id in this campaign."
          }
        },
        "tags": [
          "EntityCrud"
        ]
      }
    },
...
  },
  "info": {
    "title": "Lestapenna API",
    "description": "Read-only campaign data for the authenticated web app.",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "lp_session"
      }
    },
    "schemas": {
...
      "AskConversationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string",
            "description": "Derived from the first question, renameable."
          },
          "shared": {
            "type": "boolean",
            "description": "Published read-only to the rest of the table."
          },
          "owned": {
            "type": "boolean",
            "description": "False when another member shared it with you."
          },
          "created_at": {
            "type": "number"
          },
          "updated_at": {
            "type": "number"
          },
          "message_count": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "title",
          "shared",
          "owned",
          "created_at",
          "updated_at",
          "message_count"
        ]
      },
      "PaginatedAskConversationDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AskConversationDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Total rows matching the query, ignoring limit/offset."
          },
          "limit": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
      "AskMessageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant"
            ]
          },
          "content": {
            "type": "string"
          },
          "created_at": {
            "type": "number"
          },
          "cost_usd": {
            "type": "object",
            "nullable": true
          },
          "cost_eur": {
            "type": "object",
            "nullable": true
          },
          "provider": {
            "type": "object",
            "nullable": true
          },
          "model": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "role",
          "content",
          "created_at",
          "cost_usd",
          "cost_eur",
          "provider",
          "model"
        ]
      },
      "AskConversationDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string",
            "description": "Derived from the first question, renameable."
          },
          "shared": {
            "type": "boolean",
            "description": "Published read-only to the rest of the table."
          },
          "owned": {
            "type": "boolean",
            "description": "False when another member shared it with you."
          },
          "created_at": {
            "type": "number"
          },
          "updated_at": {
            "type": "number"
          },
          "message_count": {
            "type": "number"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AskMessageDto"
            }
          }
        },
        "required": [
          "id",
          "title",
          "shared",
          "owned",
          "created_at",
          "updated_at",
          "message_count",
          "messages"
        ]
      },
      "AskQuestionBodyDto": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string",
            "description": "The question for the Bard.",
            "maxLength": 2000
          }
        },
        "required": [
          "question"
        ]
      },
      "AskAnswerDto": {
        "type": "object",
        "properties": {
          "conversation": {
            "$ref": "#/components/schemas/AskConversationDto"
          },
          "message": {
            "description": "The Bard answer that was just persisted.",
            "allOf": [
              {
                "$ref": "#/components/schemas/AskMessageDto"
              }
            ]
          },
          "grounded": {
            "type": "object",
            "nullable": true
          },
          "missing": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "conversation",
          "message",
          "grounded",
          "missing"
        ]
      },
      "AskConversationPatchDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "New title. Empty or whitespace-only is rejected."
          },
          "shared": {
            "type": "boolean",
            "description": "Publish or withdraw the conversation for the rest of the table."
          }
        }
      },
...
      "EntityMutationResultDto": {
        "type": "object",
        "properties": {
          "short_id": {
            "type": "object",
            "description": "Five-character short id of the row written."
          }
        },
        "required": [
          "short_id"
        ]
      },
      "EntityDeleteReportDto": {
        "type": "object",
        "properties": {
          "history_deleted": {
            "type": "number",
            "description": "Righe di storico rimosse."
          },
          "rag_fragments_deleted": {
            "type": "number",
            "description": "Frammenti di memoria RAG rimossi."
          },
          "rag_refs_stripped": {
            "type": "number",
            "description": "Riferimenti tipizzati ripuliti dai frammenti superstiti."
          },
          "affiliations_deleted": {
            "type": "number",
            "description": "Affiliazioni di fazione rimosse."
          },
          "media_deleted": {
            "type": "boolean",
            "description": "True when an uploaded image existed and was removed."
          }
        },
        "required": [
          "history_deleted",
          "rag_fragments_deleted",
          "rag_refs_stripped",
          "affiliations_deleted",
          "media_deleted"
        ]
      },
      "EntityDeleteResultDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Human-readable name of the deleted entity."
          },
          "report": {
            "$ref": "#/components/schemas/EntityDeleteReportDto"
          }
        },
        "required": [
          "name",
          "report"
        ]
      },
...
    }
  }
}